+++ src/main/java/team/themoment/hellogsmv3/domain/applicant/dto/request/ApplicantReqDto.java
@@ -3,7 +3,7 @@ package team.themoment.hellogsmv3.domain.applicant.dto.request;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Pattern;
-import team.themoment.hellogsmv3.domain.applicant.type.Gender;
+import team.themoment.hellogsmv3.domain.member.entity.type.Gender;
 
 import java.time.LocalDate;
