+++ src/main/java/team/themoment/hellogsmv3/domain/applicant/service/CreateApplicantService.java
@@ -10,10 +10,8 @@ import team.themoment.hellogsmv3.domain.applicant.entity.AuthenticationCode;
 import team.themoment.hellogsmv3.domain.applicant.repo.ApplicantRepository;
 import team.themoment.hellogsmv3.domain.auth.entity.Authentication;
 import team.themoment.hellogsmv3.domain.auth.repo.AuthenticationRepository;
-import team.themoment.hellogsmv3.domain.auth.type.Role;
+import team.themoment.hellogsmv3.domain.member.entity.type.Role;
 import team.themoment.hellogsmv3.global.exception.error.ExpectedException;
-
-import java.util.List;
 
 @Service
 @Transactional
