--- src/main/java/team/themoment/hellogsmv3/domain/applicant/type/Gender.java
@@ -1,6 +1,0 @@
-package team.themoment.hellogsmv3.domain.applicant.type;
-
-public enum Gender {
-    MALE,
-    FEMALE
-}
