+++ src/main/java/team/themoment/hellogsmv3/domain/application/dto/response/GeneralAdmissionInfoResDto.java
@@ -2,7 +2,7 @@ package team.themoment.hellogsmv3.domain.application.dto.response;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
 import lombok.Builder;
-import team.themoment.hellogsmv3.domain.applicant.type.Gender;
+import team.themoment.hellogsmv3.domain.member.entity.type.Gender;
 import team.themoment.hellogsmv3.domain.application.type.DesiredMajors;
 import team.themoment.hellogsmv3.domain.application.type.GraduationStatus;
 import team.themoment.hellogsmv3.domain.application.type.Screening;
