+++ src/main/java/team/themoment/hellogsmv3/domain/auth/dto/BasicAuthenticationDto.java
@@ -3,7 +3,7 @@ package team.themoment.hellogsmv3.domain.auth.dto;
 import jakarta.persistence.EnumType;
 import jakarta.persistence.Enumerated;
 import lombok.Builder;
-import team.themoment.hellogsmv3.domain.auth.type.Role;
+import team.themoment.hellogsmv3.domain.member.entity.type.Role;
 
 @Builder
 public record BasicAuthenticationDto(
