+++ src/main/java/team/themoment/hellogsmv3/domain/auth/entity/Authentication.java
@@ -2,7 +2,7 @@ package team.themoment.hellogsmv3.domain.auth.entity;
 
 import jakarta.persistence.*;
 import lombok.*;
-import team.themoment.hellogsmv3.domain.auth.type.Role;
+import team.themoment.hellogsmv3.domain.member.entity.type.Role;
 
 @Getter
 @Entity
