+++ src/main/java/team/themoment/hellogsmv3/domain/member/entity/type/Gender.java
@@ -1,0 +1,6 @@
+package team.themoment.hellogsmv3.domain.member.entity.type;
+
+public enum Gender {
+    MALE,
+    FEMALE
+}
